+++ CHANGELOG.md
@@ -331,3 +331,4 @@
 * Make `a` paremeter in `:~.` to be existentially qualified
 * Rename type operators: `|->` to `:=>` and `<-|` to `<:=`
 * Define experimental `||=` method to apply a function inside transformer
+* Define `subview` method in `Substructure` typeclass